indexer/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.1
pypdf==5.1.0
psycopg2-binary==2.9.10
google-generativeai==0.8.3
numpy>=1.26.0,<2.0.0
python-dotenv==1.0.1
pydantic==2.10.2
pydantic-settings==2.6.1
langchain==0.3.7
langchain-text-splitters==0.3.2
nltk==3.9.1
